background-removal/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
Pillow==10.1.0
PyMatting==1.1.12
requests==2.31.0
scipy==1.10.1
sympy==1.12
torch==2.1.2
torchvision==0.16.2
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
zipp==3.17.0
